Update README.md
Suggest changing the order of updating storage with creating the signature.  As originally written a crash/power outage before updating the storage could allow for a single leaf to be used for multiple signatures.  By updating storage first the risk of a power outage is a leaf that is never used.